EdgeModules/Lego/app/requirements.txt
tensorflow==1.12.0
pillow==5.3.0
numpy==1.15.4
h5py==2.8.0
flask==1.0.2
keras==2.2.4
opencv-python==3.4.4.19